+++ qcengine/programs/tests/standard_suite_contracts.py
@@ -22,8 +22,11 @@ def query_qcvar(obj: Any, pv: str) -> Any:
             # qcel.AtomicResult.extras["qcvars"]
             vval = obj.get(pv)
             if vval is None:
-                # qcel.AtomicResult.properties
+                # qcel.AtomicResult.properties dict
                 vval = obj.get(qcvars_to_atomicproperties[pv])
+        except TypeError:
+            # qcel.AtomicResult.properties object
+            vval = getattr(obj, qcvars_to_atomicproperties[pv])
 
     return vval
 
@@ -159,8 +162,9 @@ def contractual_mp2(
             (
                 (
                     (qc_module == "gamess" and reference in ["uhf", "rohf"] and method == "mp2")
+                    or (qc_module == "gamess" and reference == "rhf" and method == "mp2" and driver == "gradient")
                     or (qc_module == "gamess" and reference in ["rhf"] and method in ["ccsd", "ccsd(t)"])
-                    or (qc_module == "nwchem-tce" and method == "mp2")
+                    or (qc_module == "nwchem-tce" and method in ["mp2", "mp3"])
                     or (qc_module == "nwchem" and reference in ["rhf"] and method in ["ccsd", "ccsd(t)"])
                     or (
                         qc_module == "psi4-occ"
@@ -249,10 +253,14 @@ def contractual_mp2p5(
         expected = True
         if (
             (
-                qc_module == "psi4-occ"
-                and reference == "rhf"
-                and corl_type in ["df", "cd"]
-                and method in ["mp2.5", "mp3"]
+                (qc_module.startswith("cfour") and method == "mp3")
+                or (
+                    qc_module == "psi4-occ"
+                    and reference == "rhf"
+                    and corl_type in ["df", "cd"]
+                    and method in ["mp2.5", "mp3"]
+                )
+                or (qc_module.startswith("nwchem") and method == "mp3")
             )
             and pv in ["MP2.5 SAME-SPIN CORRELATION ENERGY", "MP2.5 OPPOSITE-SPIN CORRELATION ENERGY"]
         ) or (
@@ -302,6 +310,7 @@ def contractual_mp3(
                     and corl_type in ["df", "cd"]
                     and method in ["mp2.5", "mp3"]
                 )
+                or (qc_module.startswith("nwchem") and method == "mp3")
             )
             and pv in ["MP3 SAME-SPIN CORRELATION ENERGY", "MP3 OPPOSITE-SPIN CORRELATION ENERGY"]
         ) or (